fix: update HEX32 to HEX24 regex for Mongo ObjectId validation and improve primary key extraction logic feat: use ObjectId for car resource creation in demo and change data type to DECIMAL in resource template

+++ dev-demo/resources/carsResourseTemplate.ts
@@ -59,7 +59,7 @@ export default function carsResourseTemplate(resourceId: string, dataSource: str
         inputSuffix: 'USD',
         allowMinMaxQuery: true,
         editingNote: 'Price is in USD',
-        type: AdminForthDataTypes.FLOAT,
+        type: AdminForthDataTypes.DECIMAL,
         required: true,
       },
       {
